# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "MiniMarket Mock API",
    "version": "2.0.0",
    "description": "쇼핑몰 QA 미니프로젝트용 Mock API (Express + JSON DB)."
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "UUID"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/api/health": {
      "get": {
        "security": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/products": {
      "get": {
        "security": [],
        "parameters": [
          {
            "name": "cat",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ship",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "min",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "max",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "rate",
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "inStock",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "perPage",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "상품 목록"
          }
        }
      }
    },
    "/api/products/{id}": {
      "get": {
        "security": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "상품 상세"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/coupons": {
      "get": {
        "security": [],
        "responses": {
          "200": {
            "description": "활성 쿠폰 목록"
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                },
                "required": [
                  "email",
                  "password"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "로그인 성공"
          },
          "401": {
            "description": "Invalid credentials"
          }
        }
      }
    },
    "/api/auth/logout": {
      "post": {
        "responses": {
          "200": {
            "description": "로그아웃"
          }
        }
      }
    },
    "/api/cart": {
      "get": {
        "responses": {
          "200": {
            "description": "장바구니 조회"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/cart/items": {
      "post": {
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "productId": {
                    "type": "string"
                  },
                  "qty": {
                    "type": "integer"
                  },
                  "opt": {
                    "type": "object",
                    "properties": {
                      "size": {
                        "type": "string"
                      },
                      "color": {
                        "type": "string"
                      }
                    }
                  }
                },
                "required": [
                  "productId"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "담기"
          },
          "409": {
            "description": "재고부족"
          }
        }
      }
    },
    "/api/cart/items/{lineId}": {
      "patch": {
        "parameters": [
          {
            "name": "lineId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "qty": {
                    "type": "integer"
                  }
                },
                "required": [
                  "qty"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "수량 변경"
          },
          "409": {
            "description": "재고부족"
          }
        }
      },
      "delete": {
        "parameters": [
          {
            "name": "lineId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "삭제"
          }
        }
      }
    },
    "/api/orders/checkout": {
      "post": {
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "address": {
                    "type": "object",
                    "properties": {
                      "zip": {
                        "type": "string"
                      },
                      "line1": {
                        "type": "string"
                      },
                      "line2": {
                        "type": "string"
                      }
                    },
                    "required": [
                      "zip",
                      "line1"
                    ]
                  },
                  "payment": {
                    "type": "object",
                    "properties": {
                      "method": {
                        "type": "string"
                      },
                      "card": {
                        "type": "string"
                      },
                      "vbank": {
                        "type": "string"
                      }
                    },
                    "required": [
                      "method"
                    ]
                  },
                  "couponCode": {
                    "type": "string"
                  },
                  "pointsUse": {
                    "type": "integer"
                  },
                  "memo": {
                    "type": "string"
                  }
                },
                "required": [
                  "address",
                  "payment"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "주문 생성"
          },
          "400": {
            "description": "검증 실패"
          },
          "409": {
            "description": "재고/품절"
          }
        }
      }
    },
    "/api/orders": {
      "get": {
        "responses": {
          "200": {
            "description": "내 주문 목록"
          }
        }
      }
    },
    "/api/orders/{id}": {
      "get": {
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "주문 상세"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/api/admin/overview": {
      "get": {
        "responses": {
          "200": {
            "description": "관리자 요약"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/api/admin/products": {
      "get": {
        "responses": {
          "200": {
            "description": "상품 전체"
          }
        }
      }
    },
    "/api/admin/products/{id}": {
      "patch": {
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "상품 수정"
          }
        }
      }
    },
    "/api/admin/orders": {
      "get": {
        "responses": {
          "200": {
            "description": "주문 전체"
          }
        }
      }
    },
    "/api/admin/orders/{id}/status": {
      "patch": {
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": {
                    "type": "string"
                  }
                },
                "required": [
                  "status"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "상태 변경"
          },
          "409": {
            "description": "전이 불가"
          }
        }
      }
    },
    "/api/admin/orders/{id}/cancel": {
      "post": {
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "취소"
          }
        }
      }
    },
    "/api/admin/orders/{id}/refund": {
      "post": {
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "환불"
          }
        }
      }
    },
    "/api/admin/audit": {
      "get": {
        "responses": {
          "200": {
            "description": "감사 로그"
          }
        }
      }
    },
    "/api/debug/reset": {
      "post": {
        "security": [],
        "responses": {
          "200": {
            "description": "초기화"
          }
        }
      }
    }
  }
}